Add dedicated header for syscall hooking
Fix redundent code: "extern ssize_t (*real_sys_write)(...);" in "logger.c"
by adding a header called "syscall_hook.h" for globalizing "real_sys_write"
